move useCase para o modulo de domain
